TVOS: Add AppleTV toolchain
Add appletv as a separate toolchain. The toolchain is very similar to
the iphone toolchain but with some modifications.
AppleTVOS doesn't support process forking which means that libraries
which have features that utilizes forking has to be built with fork
disabled.

A separate toolchain for tvOS is added to cctools-port user_examples.

A different CMake toolchain file is also used for tvOS.