WIN9X: Bring FLAC support back, through FLAC 1.2.1 (with patches)
Current FLAC expects `_stat64` to always be available on Windows, but
that's not the case for Win9x.

Moreover, even FLAC 1.3.2 in 2017 started to "Assume all currently
used OSes support SSE2", so it's getting harder and harder to have
a modern FLAC running on old Windows releases.

FLAC 1.2.1 appears to do the job, though (with some patches).
It's old (2007), but if your target is a 1995 OS, it shouldn't be
too bad.